quarkus-getting-started/src/main/java/org/acme/Pedido.java
package org.acme;

import java.util.List;

public class Pedido {

    private List<ItemPedido> itens;
    private DescontoService descontoService;

    public Pedido(List<ItemPedido> itens, DescontoService descontoService) {
        this.itens = itens;
        this.descontoService = descontoService;
    }

    public double calcularValorTotal() {
        if (itens.isEmpty()) {
            return 0.0;
        }

        double valorTotal = 0.0;
        for (ItemPedido item : itens) {
            valorTotal += item.getSubtotal();
        }

        double desconto = descontoService.calcularDesconto(valorTotal);

        if (valorTotal < desconto) {
            throw new IllegalArgumentException();
        }
        return valorTotal - desconto;
        
    }
}